+++ tests/Autocomplete.test.cpp
@@ -308,4 +308,38 @@ TEST_CASE_FIXTURE(Fixture, "get_property_changed_signal_includes_properties")
     checkStringCompletionExists(result, "Parent");
 }
 
+TEST_CASE_FIXTURE(Fixture, "get_property_changed_signal_does_not_include_children_from_sourcemap")
+{
+    loadSourcemap(R"(
+    {
+        "name": "Game",
+        "className": "DataModel",
+        "children": [
+            {
+                "name": "ReplicatedStorage",
+                "className": "ReplicatedStorage"
+            }
+        ]
+    })");
+
+    auto [source, marker] = sourceWithMarker(R"(
+        --!strict
+        game:GetPropertyChangedSignal("| ")
+    )");
+
+    auto uri = newDocument("foo.luau", source);
+
+    lsp::CompletionParams params;
+    params.textDocument = lsp::TextDocumentIdentifier{uri};
+    params.position = marker;
+
+    auto result = workspace.completion(params);
+
+    CHECK_EQ(result.size(), 3);
+    CHECK_EQ(getItem(result, "ReplicatedStorage"), std::nullopt);
+    checkStringCompletionExists(result, "Name");
+    checkStringCompletionExists(result, "Parent");
+    checkStringCompletionExists(result, "ClassName");
+}
+
 TEST_SUITE_END();