+++ CHANGELOG.md
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Fix static linking with MSVC Runtime for release binaries
 - Fixed clients that do not support pull diagnostics erroring with "server not yet received configuration for diagnostics"
+- Don't show children in autocomplete for `:GetPropertyChangedSignal("")`
 
 ## [1.31.1] - 2024-07-07